+++ src/vis_io/read/ms/tests.rs
@@ -38,7 +38,7 @@ fn test_1090008640_cross_vis() {
     let TileBaselineMaps {
         tile_to_unflagged_cross_baseline_map,
         ..
-    } = TileBaselineMaps::new(total_num_tiles, &[]);
+    } = TileBaselineMaps::new(total_num_tiles, &HashSet::new());
 
     assert_abs_diff_eq!(
         obs_context.timestamps.first().as_gpst_seconds(),
@@ -114,7 +114,7 @@ fn read_1090008640_auto_vis() {
         vis.view_mut(),
         vis_weights.view_mut(),
         *obs_context.all_timesteps.first(),
-        &[],
+        &HashSet::new(),
         &HashSet::new(),
     );
     assert!(result.is_ok(), "{}", result.unwrap_err());
@@ -191,7 +191,7 @@ fn read_1090008640_auto_vis_with_flags() {
     let obs_context = &ms_reader.obs_context;
     let total_num_tiles = obs_context.get_total_num_tiles();
     let num_chans = obs_context.num_fine_chans_per_coarse_chan;
-    let tile_flags = [1, 9];
+    let tile_flags = HashSet::from([1, 9]);
     let num_unflagged_tiles = total_num_tiles - tile_flags.len();
     let chan_flags = HashSet::from([1]);
     let num_unflagged_chans = num_chans - chan_flags.len();
@@ -292,7 +292,7 @@ fn read_1090008640_cross_and_auto_vis() {
     let TileBaselineMaps {
         tile_to_unflagged_cross_baseline_map,
         ..
-    } = TileBaselineMaps::new(total_num_tiles, &[]);
+    } = TileBaselineMaps::new(total_num_tiles, &HashSet::new());
 
     assert_abs_diff_eq!(
         obs_context.timestamps.first().as_gpst_seconds(),
@@ -315,7 +315,7 @@ fn read_1090008640_cross_and_auto_vis() {
         auto_vis_weights.view_mut(),
         *obs_context.all_timesteps.first(),
         &tile_to_unflagged_cross_baseline_map,
-        &[],
+        &HashSet::new(),
         &HashSet::new(),
     );
     assert!(result.is_ok(), "{}", result.unwrap_err());